requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
Pillow
python-dotenv